added field validation and updated float fields to BigDecimal

+++ src/main/java/com/creamsale/payload/cashback/CashBackSaleRequest.java
@@ -1,13 +1,15 @@
 package com.creamsale.payload.cashback;
 
+import java.math.BigDecimal;
+
 public class CashBackSaleRequest {
 
     private final Long cashBackId;
     private final Long shopId;
-    private final Float sale;
+    private final BigDecimal sale;
     private final String description;
 
-    public CashBackSaleRequest(Long cashBackId, Long shopId, Float sale, String description) {
+    public CashBackSaleRequest(Long cashBackId, Long shopId, BigDecimal sale, String description) {
         this.cashBackId = cashBackId;
         this.shopId = shopId;
         this.sale = sale;
@@ -22,7 +24,7 @@ public class CashBackSaleRequest {
         return shopId;
     }
 
-    public Float getSale() {
+    public BigDecimal getSale() {
         return sale;
     }
