+++ src/main/java/com/creamsale/payload/product/ProductResponse.java
@@ -2,18 +2,20 @@ package com.creamsale.payload.product;
 
 import com.creamsale.domain.Currency;
 
+import java.math.BigDecimal;
+
 public class ProductResponse {
 
     private final Long id;
     private final String name;
     private final String fullName;
     private final Long shopId;
-    private final Float price;
+    private final BigDecimal price;
     private final Currency currency;
     private final String shopCategory;
     private final String imageLink;
 
-    public ProductResponse(Long id, String name, String fullName, Long shopId, Float price, Currency currency,
+    public ProductResponse(Long id, String name, String fullName, Long shopId, BigDecimal price, Currency currency,
                            String shopCategory, String imageLink) {
         this.id = id;
         this.name = name;
@@ -41,7 +43,7 @@ public class ProductResponse {
         return shopId;
     }
 
-    public Float getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
